% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml_prcomp.R
\name{preprocess_prcomp}
\alias{preprocess_prcomp}
\title{Create preprocessing steps for prcomp objects}
\usage{
preprocess_prcomp(eigenvector, center, scale, j)
}
\arguments{
\item{eigenvector}{Eigenvector containing variable names}

\item{center}{The centering used, or FALSE}

\item{scale}{The scaling used, or FALSE}

\item{j}{Selected eigenvector index}
}
\value{
A WrapData object containing the preprocessing steps
}
\description{
Create preprocessing steps for prcomp objects
}
